# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Absen Ok API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/list-announcements": {
            "get": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Displays entire list of announcements",
                "description": "All users can see announcements made by the project manager.",
                "operationId": "App\\Http\\Controllers\\Dashboard\\Notifications\\AnnouncementController::index",
                "responses": {
                    "200": {
                        "description": "Success: Announcement data list was successfully retrieved.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Displays entire list of announcements.",
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "description": "Displays entire list of announcements.",
                                            "type": "string",
                                            "example": "Announcement Data List"
                                        },
                                        "data": {
                                            "description": "Displays entire list of announcements.",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Displays entire list of announcements.",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "picture": {
                                                        "description": "Displays entire list of announcements.",
                                                        "type": "string",
                                                        "example": "xTZCMwEpjGda3cIy4up6nMhu5urO5vufMaKjsBzP.jpg"
                                                    },
                                                    "message": {
                                                        "description": "Displays entire list of announcements.",
                                                        "type": "string",
                                                        "example": "Announcement Title"
                                                    },
                                                    "created_at": {
                                                        "description": "Displays entire list of announcements.",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-26T06:20:35.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "Displays entire list of announcements.",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-26T06:20:35.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error: Announcement Data List Not Found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Displays entire list of announcements.",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Displays entire list of announcements.",
                                            "type": "string",
                                            "example": "Announcement Data List Not Found"
                                        },
                                        "data": {
                                            "description": "Displays entire list of announcements.",
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add-announcement": {
            "post": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Save new announcements.",
                "description": "Project manager makes a new announcement for users.",
                "operationId": "App\\Http\\Controllers\\Dashboard\\Notifications\\AnnouncementController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "picture": {
                                        "description": "Supporting images for announcements.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "message": {
                                        "description": "Announcement message.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Error: Validation failed. Please check your input data!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Save new announcements.",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Save new announcements.",
                                            "type": "string",
                                            "example": "Validation failed. Please check your input data!"
                                        },
                                        "data": {
                                            "description": "Save new announcements.",
                                            "properties": {
                                                "picture": {
                                                    "description": "Pesan kesalahan untuk bidang 'picture'.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The picture field is required."
                                                    ]
                                                },
                                                "message": {
                                                    "description": "Pesan kesalahan untuk bidang 'message'.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The message field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success: Successfully added a new announcement.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Save new announcements.",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Save new announcements.",
                                            "type": "boolean",
                                            "example": "Successfully added a new announcement."
                                        },
                                        "data": {
                                            "description": "Save new announcements.",
                                            "properties": {
                                                "picture": {
                                                    "description": "URL gambar.",
                                                    "type": "string",
                                                    "example": "/storage/images/a9qy0q3d1xCkrVy55m34vxuIx9dbXcvmLmQLDwnO.jpg"
                                                },
                                                "message": {
                                                    "description": "Pesan atau informasi terkait gambar.",
                                                    "type": "string",
                                                    "example": "Announcement message."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Failed to make a new announcement!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Save new announcements.",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Save new announcements.",
                                            "type": "string",
                                            "example": "Failed to make a new announcement!"
                                        },
                                        "data": {
                                            "description": "Save new announcements.",
                                            "type": "string",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}